feat: Manage UI theme state with Recoil
