+++ app/common/ThemeButton.tsx
@@ -4,7 +4,7 @@
 import { DarkMode, LightMode } from "@mui/icons-material";
 import { IconButton, IconButtonProps } from "@mui/material";
 import { useTheme } from "@mui/material/styles";
-import { useToggleTheme } from "../core/theme.js";
+import { useToggleTheme } from "../theme/index.js";
 
 function ThemeButton(props: ThemeButtonProps): JSX.Element {
   const { ...other } = props;
